Add --purge and --no-act to kpa-backup.sh

+++ script/kpa-backup.sh
@@ -6,6 +6,8 @@ KPAUIRC=$HOME/.kde/share/apps/kphotoalbum/kphotoalbumui.rc
 BACKUP_LOCATION=~/kpa-backup
 ACTION=
 ADD_FILES_RELATIVE="exif-info.db layout.dat"
+KEEP_NUM=5
+NO_ACT=
 
 ###
 # Helper functions:
@@ -21,12 +23,16 @@ print_help()
 	echo "Usage: $0 -b|--backup [-d|--directory BACKUP_LOCATION]" >&2
 	echo "       $0 -r|--restore [-d|--directory BACKUP_LOCATION]" >&2
 	echo "       $0 -l|--list [-d|--directory BACKUP_LOCATION]" >&2
+	echo "       $0 -p|--purge [--keep NUM]" >&2
 	echo "" >&2
 	echo "Create or restore a backup of your essential KPhotoalbum files." >&2
 	echo "Note: your actual image-files are not backed up!" >&2
 	echo "" >&2
 	echo "-d|--directory BACKUP_LOCATION   Use the specified path as backup location" >&2
 	echo "                                 [default: $BACKUP_LOCATION]" >&2
+	echo "--keep NUM                       Keep the latest NUM backups" >&2
+	echo "                                 [default: $KEEP_NUM]" >&2
+	echo "-n|--no-act                      Do not take any action." >&2
 	echo "" >&2
 }
 
@@ -35,7 +41,8 @@ untar_if_changed()
 # use first parameter -p to print when the file has changed, but not untar it.
 {
 	local printonly=false
-	local diffredir="&1"
+	[ -n "$NO_ACT" ] && printonly=true
+	local diffredir="/dev/stdout"
 	local tarfile
 	for param
 	do
@@ -109,28 +116,34 @@ do_backup()
 	if [ ! -d "$BACKUP_LOCATION" ]
 	then
 		echo "Backup location ($BACKUP_LOCATION) is not a directory, creating it." >&2
-		mkdir "$BACKUP_LOCATION" || exit 1
+		[ -z "$NO_ACT" ] && mkdir "$BACKUP_LOCATION" || exit 1
 	fi
 	BACKUP_LOCATION_WDATE="$BACKUP_LOCATION"/"`date +%Y%m%d-%H%M%S`"
 	echo "Writing backup to $BACKUP_LOCATION_WDATE"
-	mkdir "$BACKUP_LOCATION_WDATE"
+	[ -z "$NO_ACT" ] && mkdir "$BACKUP_LOCATION_WDATE"
 	if [ -e "$BACKUP_LOCATION"/latest ]
 	then
-		rm "$BACKUP_LOCATION"/latest
+		[ -z "$NO_ACT" ] && rm "$BACKUP_LOCATION"/latest
 	fi
-	ln -s "$BACKUP_LOCATION_WDATE" "$BACKUP_LOCATION"/latest
+	[ -z "$NO_ACT" ] && ln -s "$BACKUP_LOCATION_WDATE" "$BACKUP_LOCATION"/latest
 
 	echo "Backing up essential files..."
-	for f in "$KPARC" "$INDEXFILE"
-	do
-		tar -Pcvz -f "$BACKUP_LOCATION_WDATE"/`basename "$f"`.tgz "$f"
-	done
+	if [ -z "$NO_ACT" ]
+	then
+		for f in "$KPARC" "$INDEXFILE"
+		do
+			tar -Pcvz -f "$BACKUP_LOCATION_WDATE"/`basename "$f"`.tgz "$f"
+		done
+	fi
 	echo "Backing up additional files..."
-	[ -f "$KPAUIRC" ] && tar -Pcvz -f "$BACKUP_LOCATION_WDATE"/`basename "$KPAUIRC"`.tgz "$KPAUIRC"
-	for f in $ADD_FILES_RELATIVE
-	do
-		[ -f "$KPA_FOLDER/$f" ] && tar -Pcvz -f "$BACKUP_LOCATION_WDATE"/`basename "$f"`.tgz "$KPA_FOLDER/$f"
-	done
+	if [ -z "$NO_ACT" ]
+	then
+		[ -f "$KPAUIRC" ] && tar -Pcvz -f "$BACKUP_LOCATION_WDATE"/`basename "$KPAUIRC"`.tgz "$KPAUIRC"
+		for f in $ADD_FILES_RELATIVE
+		do
+			[ -f "$KPA_FOLDER/$f" ] && tar -Pcvz -f "$BACKUP_LOCATION_WDATE"/`basename "$f"`.tgz "$KPA_FOLDER/$f"
+		done
+	fi
 }
 
 do_restore()
@@ -138,6 +151,7 @@ do_restore()
 	echo "Restoring essential files..."
 	for f in "$BACKUP_LOCATION/latest"/*.tgz
 	do
+		# untar_if_changed honors NO_ACT:
 		untar_if_changed "$f"
 	done
 }
@@ -169,11 +183,33 @@ do_list()
 	done
 }
 
+do_purge()
+{
+	cd "$BACKUP_LOCATION"
+	# list newest entries first, skip KEEP_NUM directories:
+	for d in `ls -t1`
+	do
+		if [ -h "$d" -o ! -d "$d" ]
+		then # skip "latest"
+			echo "Skipping non-directory $d"
+			continue
+		fi
+		if [ "$KEEP_NUM" -gt 0 ]
+		then
+			echo "Skipping $d..."
+			let KEEP_NUM--
+			continue
+		fi
+		echo "Purging backup $d..."
+		[ -z "$NO_ACT" ] && rm -rf "$d"
+	done
+}
+
 ###
 # Parse commandline:
 ###
 
-TEMP=`getopt -o hbrld: --long help,backup,restore,list,directory: \
+TEMP=`getopt -o hbrlpnd: --long help,backup,restore,list,purge,no-act,directory:,keep: \
      -n 'kpa-backup' -- "$@"`
 
 if [ $? != 0 ] ; then echo "Terminating..." >&2 ; exit 1 ; fi
@@ -187,7 +223,10 @@ while true ; do
 		-b|--backup) ACTION=do_backup ; shift ;;
 		-r|--restore) ACTION=do_restore ; shift ;; 
 		-l|--list) ACTION=do_list ; shift ;; 
+		-p|--purge) ACTION=do_purge ; shift ;;
+		-n|--no-act) NO_ACT=1 ; shift ;;
 		-d|--directory) BACKUP_LOCATION="$2" ; shift 2 ;;
+		--keep) KEEP_NUM="$2" ; shift 2 ;;
 		--) shift ; break ;;
 		*) echo "Internal error!" ; exit 1 ;;
 	esac
